Creado el esqueleto de las clases básicas del MVC y modificado un pequeño detalle del diagrama de clases
