+++ src/Aplicacion/Controller.java
@@ -1,0 +1,40 @@
+package Aplicacion;
+
+/**
+ * Controlador de la aplicación
+ */
+public class Controller {
+    
+    /**
+     * Accede al listado de coches
+     * @param form Filtro de parámetros
+     */
+    public static void list(Form form){
+        
+    }
+    
+    /**
+     * Accede a la ficha de un coche
+     * @param carName Nombre identificador del coche
+     */
+    public static void view(String carName){
+        
+    }
+    
+    /**
+     * Accede a la pantalla de edición de un coche
+     * @param carName Nombre identificador del coche
+     */
+    public static void edit(String carName){
+        
+    }
+    
+    /**
+     * Envía una solicitud de contacto para concertar una cita y ver un coche
+     * @param form Filtro de parámetros
+     * @param carName Nombre identificador del coche que desea ver el cliente
+     */
+    public static void contactRequest(Form form, String carName){
+        
+    }
+}
