+++ src/Aplicacion/View.java
@@ -1,0 +1,36 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package Aplicacion;
+
+/**
+ * Vista de la aplicación
+ */
+public class View {
+    
+    /**
+     * Visualiza una lista de coches
+     * @param models Lista de coches
+     */
+    public static void list(Car_Model[] models){
+        
+    }
+    
+    /**
+     * Muestra la ficha de un coche
+     * @param models Coche a mostrar
+     */
+    public static void view(Car_Model models){
+        
+    }
+    
+    /**
+     * Muestra la pantalla de edición de un coche
+     * @param model Coche a editar
+     */
+    public static void edit(Car_Model model){
+        
+    }
+}
